requirements.txt
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
chardet==4.0.0
click==8.1.7
cssselect==1.2.0
cssutils==2.9.0
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.110.0
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.3
json5==0.9.5
jsonschema==3.2.0
lxml==5.1.0
MarkupSafe==2.1.5
orjson==3.9.15
phonenumbers==8.13.20
premailer==3.10.0
prettytable==3.9.0
pydantic==2.6.4
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
pyrsistent==0.20.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
ujson==5.9.0
urllib3==1.26.18
uvicorn==0.28.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
SQLAlchemy==2.0.29
psycopg2==2.9.9
resend==1.0.2
gunicorn
rq
redis
apscheduler==3.10.4
simple-file-checksum==1.2.2
google-auth==2.32.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.138.0
passlib==1.7.4
python-jose==3.3.0